+++ networking/p2p/src/test/java/tech/pegasys/teku/networking/p2p/network/config/NetworkConfigTest.java
@@ -14,12 +14,9 @@
 package tech.pegasys.teku.networking.p2p.network.config;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import java.nio.file.Path;
 import java.util.Optional;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
 
 @SuppressWarnings("AddressSelection")
 class NetworkConfigTest {
@@ -58,36 +55,6 @@ class NetworkConfigTest {
     final String result = createConfig().getAdvertisedIp();
     assertThat(result).isNotEqualTo("::0");
     assertThat(result).isNotEqualTo("0.0.0.0");
-  }
-
-  @Test
-  void shouldThrowExceptionIfInvalidFileName(@TempDir Path tempDir) {
-    final NetworkConfig config =
-        NetworkConfig.builder()
-            .advertisedIp(advertisedIp)
-            .networkInterface(listenIp)
-            .privateKeyFile(tempDir + "/invalid file name!!\0")
-            .build();
-
-    assertThat(config.getPrivateKeySource()).isPresent();
-    assertThatThrownBy(() -> config.getPrivateKeySource().get().getOrGeneratePrivateKeyBytes())
-        .isInstanceOf(IllegalArgumentException.class)
-        .hasMessageContaining("Not able to create or retrieve p2p private key file -");
-  }
-
-  @Test
-  void shouldThrowExceptionIfProvideDirectory(@TempDir Path tempDir) {
-    final NetworkConfig config =
-        NetworkConfig.builder()
-            .advertisedIp(advertisedIp)
-            .networkInterface(listenIp)
-            .privateKeyFile(tempDir.toString()) // directory provided instead of file
-            .build();
-
-    assertThat(config.getPrivateKeySource()).isPresent();
-    assertThatThrownBy(() -> config.getPrivateKeySource().get().getOrGeneratePrivateKeyBytes())
-        .isInstanceOf(RuntimeException.class)
-        .hasMessageContaining("p2p private key file not found -");
   }
 
   private NetworkConfig createConfig() {
